Relaciones.

+++ spotify-service/src/main/java/com/spotifyservice/spotifyservice/service/AlbumService.java
@@ -3,6 +3,7 @@ package com.spotifyservice.spotifyservice.service;
 import com.spotifyservice.spotifyservice.controller.request.AlbumRequest;
 import com.spotifyservice.spotifyservice.domain.Album;
 import com.spotifyservice.spotifyservice.domain.AlbumMapper;
+import com.spotifyservice.spotifyservice.domain.Artist;
 import com.spotifyservice.spotifyservice.repository.AlbumRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,23 +19,29 @@ public class AlbumService {
     List<Album> listaAlbums = new ArrayList<Album>();
 
     @Autowired
+    private ArtistService artistService;
+
+    @Autowired
     private AlbumMapper albumMapper;
 
     @Autowired
     private AlbumRepository albumRepository;
 
-    public void initAlbum(){
+    public void initAlbum(Artist artist){
         if (listaAlbums.isEmpty()) {
             AlbumRequest album1 = new AlbumRequest();
-            album1.setIdArtist(1L);
+            album1.setIdArtist(artist);
             album1.setName("Album1");
-            createAlbum(album1);
+            createPrimerAlbum(album1);
 
             listaAlbums.stream().forEach(album -> {
                 albumRepository.save(album);
             });
         }
     }
+
+    public Album getPrimerAlbum(){return albumRepository.findByIdAlbum(1L);}
+
     public Album getAlbum(Long id){
         return albumRepository.findByIdAlbum(id);
     }
@@ -43,10 +50,23 @@ public class AlbumService {
         return albumRepository.findAll();
     }
 
-    public Album createAlbum(AlbumRequest request){
+    public Album createPrimerAlbum(AlbumRequest request){
         Album album = albumMapper.apply(request);
         albumRepository.save(albumMapper.apply(request));
         return album;
+    }
+
+    public Album createAlbum(AlbumRequest albumRequest){
+        Album albumNew = null;
+        Album album = albumMapper.apply(albumRequest);
+        for(Artist artist: artistService.getArtists()){
+            if(artist.getIdArtist().equals(albumRequest.getIdArtist().getIdArtist())){
+                albumNew = album;
+                albumNew.setIdArtist(artist);
+                albumRepository.save(albumNew);
+            }
+        }
+        return albumNew;
     }
 
     public Album editAlbum(Long id, AlbumRequest albumRequest){
@@ -59,21 +79,36 @@ public class AlbumService {
         }
 
         if(albumRequest.getName() != null){albumActualizado.setNameAlbum(albumRequest.getName());}
-        if(albumRequest.getIdArtist() != null){albumActualizado.setIdArtist(albumRequest.getIdArtist());}
+        for(Artist artist: artistService.getArtists()){
+            if(artist.getIdArtist().equals(albumRequest.getIdArtist().getIdArtist())){
+                albumActualizado.setIdArtist(artist);
+            }
+        }
 
-        
         albumRepository.save(albumActualizado);
 
         return albumActualizado;
     }
 
-
     public Album deleteAlbum(Long id){
-        albumRepository.deleteById(id);
+        Artist art = null;
+        for(Artist artist: artistService.getArtists()){
+            if(artist.getIdArtist().equals(albumRepository.findByIdAlbum(id).getIdArtist().getIdArtist())){
+                art = artist;
+                albumRepository.deleteById(id);
+                artistService.guardarArtist(art);
+            }
+        }
         return null;
     }
 
-
+    public void setArtistNull(Long idArtist){
+        for(Album album:albumRepository.findAll()){
+            if(album.getIdArtist().getIdArtist().equals(idArtist)){
+                album.setIdArtist(null);
+            }
+        }
+    }
 
     /**
      public List<Album> getAlbum(Long id){
